+++ vck-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/openid/DummyCredentialDataProvider.kt
@@ -50,6 +50,7 @@ object DummyCredentialDataProvider {
                     expiration = expiration,
                     scheme = credentialScheme,
                     subjectPublicKey = subjectPublicKey,
+                    userInfo = DummyOAuth2DataProvider.user,
                 )
 
                 PLAIN_JWT -> CredentialToBeIssued.VcJwt(
@@ -61,6 +62,7 @@ object DummyCredentialDataProvider {
                     expiration = expiration,
                     scheme = credentialScheme,
                     subjectPublicKey = subjectPublicKey,
+                    userInfo = DummyOAuth2DataProvider.user,
                 )
 
                 ISO_MDOC -> CredentialToBeIssued.Iso(
@@ -70,6 +72,7 @@ object DummyCredentialDataProvider {
                     expiration = expiration,
                     scheme = credentialScheme,
                     subjectPublicKey = subjectPublicKey,
+                    userInfo = DummyOAuth2DataProvider.user,
                 )
             }
         } else if (credentialScheme == MobileDrivingLicenceScheme) {
@@ -102,6 +105,7 @@ object DummyCredentialDataProvider {
                 expiration = expiration,
                 scheme = credentialScheme,
                 subjectPublicKey = subjectPublicKey,
+                userInfo = DummyOAuth2DataProvider.user,
             )
         } else if (credentialScheme == EuPidScheme) {
             val subjectId = subjectPublicKey.didEncoded
@@ -155,6 +159,7 @@ object DummyCredentialDataProvider {
                         expiration = expiration,
                         scheme = credentialScheme,
                         subjectPublicKey = subjectPublicKey,
+                        userInfo = DummyOAuth2DataProvider.user,
                     )
 
                 PLAIN_JWT -> CredentialToBeIssued.VcJwt(
@@ -172,6 +177,7 @@ object DummyCredentialDataProvider {
                     expiration = expiration,
                     scheme = credentialScheme,
                     subjectPublicKey = subjectPublicKey,
+                    userInfo = DummyOAuth2DataProvider.user,
                 )
 
                 ISO_MDOC -> CredentialToBeIssued.Iso(
@@ -181,6 +187,7 @@ object DummyCredentialDataProvider {
                     expiration = expiration,
                     scheme = credentialScheme,
                     subjectPublicKey = subjectPublicKey,
+                    userInfo = DummyOAuth2DataProvider.user,
                 )
             }
         } else {