+++ vck-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/openid/DummyOAuth2IssuerCredentialDataProvider.kt
@@ -87,14 +87,16 @@ object DummyOAuth2IssuerCredentialDataProvider : CredentialDataProviderFun {
                 claims,
                 expiration,
                 ConstantIndex.AtomicAttribute2023,
-                subjectPublicKey
+                subjectPublicKey,
+                DummyOAuth2DataProvider.user,
             )
 
             PLAIN_JWT -> CredentialToBeIssued.VcJwt(
                 AtomicAttribute2023(subjectId, GIVEN_NAME, givenName ?: "no value"),
                 expiration,
                 ConstantIndex.AtomicAttribute2023,
-                subjectPublicKey
+                subjectPublicKey,
+                DummyOAuth2DataProvider.user,
             )
 
             ISO_MDOC -> CredentialToBeIssued.Iso(
@@ -103,7 +105,8 @@ object DummyOAuth2IssuerCredentialDataProvider : CredentialDataProviderFun {
                 },
                 expiration,
                 ConstantIndex.AtomicAttribute2023,
-                subjectPublicKey
+                subjectPublicKey,
+                DummyOAuth2DataProvider.user,
             )
         }
     }
@@ -124,7 +127,13 @@ object DummyOAuth2IssuerCredentialDataProvider : CredentialDataProviderFun {
             issuerSignedItem(ISSUE_DATE, "2023-01-01", digestId++),
             issuerSignedItem(EXPIRY_DATE, "2033-01-01", digestId++),
         )
-        return CredentialToBeIssued.Iso(issuerSignedItems, expiration, MobileDrivingLicenceScheme, subjectPublicKey)
+        return CredentialToBeIssued.Iso(
+            issuerSignedItems,
+            expiration,
+            MobileDrivingLicenceScheme,
+            subjectPublicKey,
+            DummyOAuth2DataProvider.user,
+        )
     }
 
     private fun getEuPid(
@@ -156,6 +165,7 @@ object DummyOAuth2IssuerCredentialDataProvider : CredentialDataProviderFun {
                 expiration,
                 EuPidScheme,
                 subjectPublicKey,
+                DummyOAuth2DataProvider.user,
             )
 
             PLAIN_JWT -> CredentialToBeIssued.VcJwt(
@@ -172,6 +182,7 @@ object DummyOAuth2IssuerCredentialDataProvider : CredentialDataProviderFun {
                 expiration,
                 EuPidScheme,
                 subjectPublicKey,
+                DummyOAuth2DataProvider.user,
             )
 
             ISO_MDOC -> CredentialToBeIssued.Iso(
@@ -181,6 +192,7 @@ object DummyOAuth2IssuerCredentialDataProvider : CredentialDataProviderFun {
                 expiration,
                 EuPidScheme,
                 subjectPublicKey,
+                DummyOAuth2DataProvider.user,
             )
         }
     }