+++ vck/src/commonTest/kotlin/at/asitplus/wallet/lib/agent/AgentRevocationTest.kt
@@ -1,5 +1,7 @@
 package at.asitplus.wallet.lib.agent
 
+import at.asitplus.openid.OidcUserInfo
+import at.asitplus.openid.OidcUserInfoExtended
 import at.asitplus.wallet.lib.agent.FixedTimePeriodProvider.timePeriod
 import at.asitplus.wallet.lib.cbor.VerifyCoseSignature
 import at.asitplus.wallet.lib.data.AtomicAttribute2023
@@ -11,7 +13,6 @@ import at.asitplus.wallet.lib.data.rfc.tokenStatusList.StatusListTokenPayload
 import at.asitplus.wallet.lib.data.rfc.tokenStatusList.agents.communication.primitives.StatusListTokenMediaType
 import at.asitplus.wallet.lib.data.rfc.tokenStatusList.primitives.TokenStatus
 import at.asitplus.wallet.lib.jws.VerifyJwsObject
-import com.benasher44.uuid.uuid4
 import io.kotest.assertions.fail
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.nulls.shouldNotBeNull
@@ -188,7 +189,8 @@ private suspend fun IssuerCredentialStore.revokeCredentialsWithIndexes(revokedIn
                 subject = cred,
                 expiration = expirationDate,
                 scheme = ConstantIndex.AtomicAttribute2023,
-                subjectPublicKey = EphemeralKeyWithoutCert().publicKey
+                subjectPublicKey = EphemeralKeyWithoutCert().publicKey,
+                userInfo = OidcUserInfoExtended.fromOidcUserInfo(OidcUserInfo("subject")).getOrThrow(),
             ),
             timePeriod
         ).getOrThrow()
@@ -210,7 +212,8 @@ private suspend fun IssuerCredentialStore.revokeRandomCredentials(): List<ULong>
                 subject = cred,
                 expiration = expirationDate,
                 scheme = ConstantIndex.AtomicAttribute2023,
-                subjectPublicKey = EphemeralKeyWithoutCert().publicKey
+                subjectPublicKey = EphemeralKeyWithoutCert().publicKey,
+                userInfo = OidcUserInfoExtended.fromOidcUserInfo(OidcUserInfo("subject")).getOrThrow(),
             ),
             timePeriod
         ).getOrThrow().statusListIndex